HBX-1992: HBX-1992: Reorganize the reverse engineering strategy hierarchy - Rename 'org.hibernate.tool.internal.reveng.TableSelectorStrategy' to 'org.hibernate.tool.internal.reveng.strategy.TableSelectorStrategy'
